Create Inventory Management System
The goal in the task is to associate items to different warehouses. Each of the items and 
warehouses has a textual name and a positive numerical code. The code is unique for each 
item and each warehouse (the code may be the same for the item and the warehouse). 
There may be several different warehouses that store the same item (same name and same 
code). Conversely, there may be several different items stored in the same warehouse.